+++ CHANGELOG.rst
@@ -26,6 +26,8 @@ Changes:
 
 - ``structlog.threadlocal.wrap_dict()`` now has a correct type annotation.
   `#290 <https://github.com/hynek/structlog/pull/290>`_
+- Fix isolation in ``structlog.contextvars``.
+  `#302 <https://github.com/hynek/structlog/pull/302>`_
 
 
 ----